[feat] 이메일 인증 로직 구현(추후 디테일 수정) 및 디렉토리 구조 변경

+++ src/main/java/com/honeyboard/api/user/model/User.java
@@ -17,11 +17,11 @@ import lombok.ToString;
 @Builder
 public class User {
 
-	private long userId;
+	private int userId;
 	private String email;
 	private String password;
 	private String name;
-	private long generationId;
+	private int generationId;
 	private String role;
 	private String loginType;
 	private boolean isSsafy;
